feat(code refactoring): code refactoring
